added all changes from backup branch
