+++ controller/admin.controller.js
@@ -1,9 +1,10 @@
-const{clearSessionValue} = require('../lib/common');
+const{clearSessionValue, mongoSanitize} = require('../lib/common');
+const bcrypt = require('bcryptjs');
 // Admin section
 const adminDashboard=( (req, res, next)=>{
     res.redirect('/admin/dashboard');
 })
-// logout
+// logout 1
 const logout=( (req, res)=>{
     req.session.user = null;
     req.session.message = null;
@@ -11,7 +12,7 @@ const logout=( (req, res)=>{
     res.redirect('/');
 })
 
-// login form
+// login form 2
 const login=(async (req, res) => {
     const db = req.app.db;
 
@@ -36,5 +37,87 @@ const login=(async (req, res) => {
     }
 });
 
+// login the user and check the password 3
+const loginValidate=( async (req, res) => {
+    const db = req.app.db;
 
-module.exports={adminDashboard, logout, login}
+    const user = await db.users.findOne({ userEmail: mongoSanitize(req.body.email) });
+    if(!user || user === null){
+        res.status(400).json({ message: 'A user with that email does not exist.' });
+        return;
+    }
+
+    // we have a user under that email so we compare the password
+    bcrypt.compare(req.body.password, user.userPassword)
+        .then((result) => {
+            if(result){
+                req.session.user = req.body.email;
+                req.session.usersName = user.usersName;
+                req.session.userId = user._id.toString();
+                req.session.isAdmin = user.isAdmin;
+                res.status(200).json({ message: 'Login successful' });
+                return;
+            }
+            // password is not correct
+            res.status(400).json({ message: 'Access denied. Check password and try again.' });
+        });
+});
+
+
+// setup form is shown when there are no users setup in the DB 4
+const adminSetup=(async (req, res) => {
+    const db = req.app.db;
+
+    const userCount = await db.users.countDocuments({});
+    // dont allow the user to "re-setup" if a user exists.
+    // set needsSetup to false as a user exists
+    req.session.needsSetup = false;
+    if(userCount === 0){
+        req.session.needsSetup = true;
+        res.render('setup', {
+            title: 'Setup',
+            config: req.app.config,
+            helpers: req.handlebars.helpers,
+            message: clearSessionValue(req.session, 'message'),
+            messageType: clearSessionValue(req.session, 'messageType'),
+            showFooter: 'showFooter'
+        });
+        return;
+    }
+    res.redirect('/admin/login');
+});
+
+
+
+// insert a user 5
+const setupUser=(async (req, res) => {   //can convert this to function because it seems the flow of control not entring try catch
+    const db = req.app.db;
+
+    const doc = {
+        usersName: req.body.usersName,
+        userEmail: req.body.userEmail,
+        userPassword: bcrypt.hashSync(req.body.userPassword, 10),
+        isAdmin: true,
+        isOwner: true
+    };
+
+    // check for users 
+    const userCount = await db.users.countDocuments({});
+    if(userCount === 0){
+        // email is ok to be used.
+        try{
+            await db.users.insertOne(doc);
+            res.status(200).json({ message: 'User account inserted' });
+            return;
+        }catch(ex){
+            console.error(colors.red(`Failed to insert user: ${ex}`));
+            res.status(200).json({ message: 'Setup failed' });
+            return;
+        }
+    }
+    res.status(200).json({ message: 'Already setup.' });
+});
+
+
+
+module.exports={adminDashboard, logout, login, loginValidate, adminSetup, setupUser}
